🚸 (SUN-71): Team Member page modifications
Team Member list no longer links to an individual team member page. We don't have enough content to warrant that just yet. Instead, the bio and social links are in line within the list. When we get to the point where there's more information to warrant a devoted page, then we can build that in.